Refactored Collection QuickOptions
